Revert "Restructuring client-gen (#88)"
This reverts commit 3bd6073cc1b7f64b9d678a4cb3b9960a3e8e86d3.
